# public/doc/swagger.yaml
swagger: '2.0'
info:
  description: MasterMind REST API ドキュメント
  version: 0.0.1
  title: MasterMind REST APIs
#host: dotnsf-mastermind.herokuapp.com
host: localhost:8080
basePath: /api
tags:
  - name: play
    description: MasterMind API for Playing
  - name: admin
    description: MasterMind API for Admin
schemes:
  - http
  - https
paths:
  /ping:
    get:
      tags:
        - admin
      summary: ping
      description: サーバーの死活監視用
      produces:
        - application/json
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /init:
    get:
      tags:
        - play
      summary: 初期化
      description: ゲーム初期化
      produces:
        - application/json
      parameters:
        - name: length
          type: integer
          in: query
          description: 数値桁数（デフォルト=4）
        - name: highlow
          type: integer
          in: query
          description: チェック時の返答に High or Low を加える（デフォルト=0）
        - name: name
          type: string
          in: query
          description: プレイヤー名（デフォルト=''）
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /guess:
    get:
      tags:
        - play
      summary: 推測
      description: 推測値に対する結果確認
      produces:
        - application/json
      parameters:
        - name: id
          type: string
          in: query
          description: 初期化時に取得した id
          required: true
        - name: value
          type: string
          in: query
          description: 推測した数値文字列
          required: true
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /giveup:
    get:
      tags:
        - play
      summary: ギブアップ
      description: ギブアップ
      produces:
        - application/json
      parameters:
        - name: id
          type: string
          in: query
          description: 初期化時に取得した id
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
  /status:
    get:
      tags:
        - admin
      summary: ステータス参照
      description: ステータス参照
      produces:
        - application/json
      parameters:
        - name: id
          type: string
          in: query
          description: 初期化時に取得した id
        - name: password
          type: string
          in: query
          description: 管理用パスワード（id パラメータ無しで実行する場合は必要）
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
        '403':
          description: パスワードエラー
  /reset:
    get:
      tags:
        - admin
      summary: 全データリセット
      description: 全データリセット
      produces:
        - application/json
      parameters:
        - name: password
          type: string
          in: query
          description: 管理用パスワード
      responses:
        '200':
          description: 成功
        '400':
          description: エラー
        '403':
          description: パスワードエラー